move strong lensing code and cataloging code around
