+++ bliss/simulator/galsim_decoder.py
@@ -187,143 +187,6 @@ class SingleGalsimGalaxyDecoder(PSFDecoder):
         return torch.from_numpy(image).reshape(1, slen, slen)
 
 
-class SingleLensedGalsimGalaxyDecoder(SingleGalsimGalaxyDecoder):
-    def __init__(
-        self,
-        slen: int,
-        n_bands: int,
-        pixel_scale: float,
-        psf_params_file: Optional[str] = None,
-        psf_slen: Optional[int] = None,
-        sdss_bands: Optional[Tuple[int, ...]] = None,
-    ) -> None:
-        super().__init__(
-            slen=slen,
-            n_bands=n_bands,
-            pixel_scale=pixel_scale,
-            psf_params_file=psf_params_file,
-            psf_slen=psf_slen,
-            sdss_bands=sdss_bands,
-        )
-
-    def __call__(
-        self,
-        z_lens: Tensor,
-        offset: Optional[Tensor] = None,
-    ) -> Tensor:
-        if z_lens.shape[0] == 0:
-            return torch.zeros(0, 1, self.slen, self.slen, device=z_lens.device)
-
-        if z_lens.shape == (12,):
-            assert offset is None or offset.shape == (2,)
-            return self.render_lensed_galaxy(z_lens, self.psf_galsim, self.slen, offset)
-
-        images = []
-        for ii, lens_params in enumerate(z_lens):
-            off = offset if not offset else offset[ii]
-            assert off is None or off.shape == (2,)
-            image = self.render_lensed_galaxy(lens_params, self.psf_galsim, self.slen, off)
-            images.append(image)
-        return torch.stack(images, dim=0).to(z_lens.device)
-
-    def sie_deflection(self, x, y, lens_params):
-        """Get deflection for grid_sample (in pixels) due to a gravitational lens.
-
-        Adopted from: Adam S. Bolton, U of Utah, 2009
-
-        Args:
-            x: images of x coordinates
-            y: images of y coordinates
-            lens_params: vector of parameters with 5 elements, defined as follows:
-                par[0]: lens strength, or 'Einstein radius'
-                par[1]: x-center
-                par[2]: y-center
-                par[3]: e1 ellipticity
-                par[4]: e2 ellipticity
-
-        Returns:
-            Tuple (xg, yg) of gradients at the positions (x, y)
-        """
-        b, center_x, center_y, e1, e2 = lens_params.cpu().numpy()
-        ell = np.sqrt(e1**2 + e2**2)
-        q = (1 - ell) / (1 + ell)
-        phirad = np.arctan(e2 / e1)
-
-        # Go into shifted coordinats of the potential:
-        xsie = (x - center_x) * np.cos(phirad) + (y - center_y) * np.sin(phirad)
-        ysie = (y - center_y) * np.cos(phirad) - (x - center_x) * np.sin(phirad)
-
-        # Compute potential gradient in the transformed system:
-        r_ell = np.sqrt(q * xsie**2 + ysie**2 / q)
-        qfact = np.sqrt(1.0 / q - q)
-
-        # (r_ell == 0) terms prevent divide-by-zero problems
-        eps = 0.001
-        if qfact >= eps:
-            xtg = (b / qfact) * np.arctan(qfact * xsie / (r_ell + (r_ell == 0)))
-            ytg = (b / qfact) * np.arctanh(qfact * ysie / (r_ell + (r_ell == 0)))
-        else:
-            xtg = b * xsie / (r_ell + (r_ell == 0))
-            ytg = b * ysie / (r_ell + (r_ell == 0))
-
-        # Transform back to un-rotated system:
-        xg = xtg * np.cos(phirad) - ytg * np.sin(phirad)
-        yg = ytg * np.cos(phirad) + xtg * np.sin(phirad)
-        return (xg, yg)
-
-    def bilinear_interpolate_numpy(self, im, x, y):
-        x0 = np.floor(x).astype(int)
-        x1 = x0 + 1
-        y0 = np.floor(y).astype(int)
-        y1 = y0 + 1
-
-        x0 = np.clip(x0, 0, im.shape[1] - 1)
-        x1 = np.clip(x1, 0, im.shape[1] - 1)
-        y0 = np.clip(y0, 0, im.shape[0] - 1)
-        y1 = np.clip(y1, 0, im.shape[0] - 1)
-
-        i_a = im[y0, x0]
-        i_b = im[y1, x0]
-        i_c = im[y0, x1]
-        i_d = im[y1, x1]
-
-        wa = (x1 - x) * (y1 - y)
-        wb = (x1 - x) * (y - y0)
-        wc = (x - x0) * (y1 - y)
-        wd = (x - x0) * (y - y0)
-
-        return (i_a.T * wa).T + (i_b.T * wb).T + (i_c.T * wc).T + (i_d.T * wd).T
-
-    def lens_galsim(self, unlensed_image, lens_params):
-        nx, ny = unlensed_image.shape
-        x_range = [-nx // 2, nx // 2]
-        y_range = [-ny // 2, ny // 2]
-        x = (x_range[1] - x_range[0]) * np.outer(np.ones(ny), np.arange(nx)) / float(
-            nx - 1
-        ) + x_range[0]
-        y = (y_range[1] - y_range[0]) * np.outer(np.arange(ny), np.ones(nx)) / float(
-            ny - 1
-        ) + y_range[0]
-
-        (xg, yg) = self.sie_deflection(x, y, lens_params)
-        lensed_image = self.bilinear_interpolate_numpy(
-            unlensed_image, (x - xg) + nx // 2, (y - yg) + ny // 2
-        )
-        return lensed_image.astype(unlensed_image.dtype)
-
-    def render_lensed_galaxy(
-        self,
-        lens_params: Tensor,
-        psf: galsim.GSObject,
-        slen: int,
-        offset: Optional[Tensor] = None,
-    ) -> Tensor:
-        lensed_galaxy_params, pure_lens_params = lens_params[:7], lens_params[7:]
-        unlensed_src = self._render_galaxy_np(lensed_galaxy_params, psf, slen, offset)
-        lensed_src = self.lens_galsim(unlensed_src, pure_lens_params)
-        return torch.from_numpy(lensed_src).reshape(1, slen, slen)
-
-
 class DefaultGalsimPrior:
     def __init__(
         self,
