+++ tests/test_catalogs.py
@@ -19,7 +19,7 @@ def test_multiple_sources_one_tile():
     # should return only first source in first tile.
     tile_cat = full_cat.to_tile_params(1, 1, ignore_extra_sources=True)
     assert torch.equal(tile_cat.n_sources, torch.tensor([[1, 0], [0, 0]]).reshape(1, 2, 2))
-    # TODO: would torch.all(torch.is_close) be better here?
+    # TODO: torch.all(torch.is_close) seems better here
     assert torch.equal(
         tile_cat.locs, torch.tensor([[[0.5, 0.5], [0, 0]], [[0, 0], [0, 0]]]).reshape(1, 2, 2, 1, 2)
     )
